+++ src/main/groovy/com/cedarsoftware/ncube/NCube.groovy
@@ -2255,17 +2255,25 @@ class NCube<T>
         clearSha1()
     }
 
-    void createNewAxisReference(final String axisName, final ApplicationID refAppId, final String refCubeName, final String refAxisName)
+    protected void convertAxisToRefAxis(final String axisName, final ApplicationID refAppId, final String refCubeName, final String refAxisName)
     {
         Axis axis = getAxis(axisName)
-        axis.createReference(refAppId, refCubeName, refAxisName)
+        if (axis.reference)
+        {
+            return
+        }
+        axis.makeReference(refAppId, refCubeName, refAxisName)
         clearSha1()
     }
 
-    void createExistingAxisReference(final String axisName, final ApplicationID refAppId, final String refCubeName, final String refAxisName)
+    protected void convertExistingAxisToRefAxis(final String axisName, final ApplicationID refAppId, final String refCubeName, final String refAxisName)
     {
         // copy list of columns before axis changes
         Axis axis = getAxis(axisName)
+        if (axis.reference)
+        {
+            return
+        }
         List<Column> oldColumns = axis.columns
 
         // make copy of the cell map to reference after the axis changes
@@ -2281,33 +2289,61 @@ class NCube<T>
         args[ReferenceAxisLoader.REF_AXIS_NAME] = refAxisName    // axis name of the referring axis (the variable that you had missing earlier)
         ReferenceAxisLoader refAxisLoader = new ReferenceAxisLoader(name, axisName, args)
         Axis newAxis = new Axis(axisName, axis.id, axis.hasDefaultColumn(), refAxisLoader)
+
+        Map<Long, Long> oldToNewId = [:]
+        for (Column oldCol : oldColumns)
+        {   // Locate columns in O(1) to O(log n)
+            if (newAxis.type == AxisType.RULE)
+            {   // Rule columns are located by ID or rule name ('name' meta-property)
+                oldToNewId[oldCol.id] = newAxis.findColumn(oldCol.columnName).id
+            }
+            else
+            {   // Use value that exists on OLD column to locate NEW column
+                oldToNewId[oldCol.id] = newAxis.findColumn(oldCol.valueThatMatches).id
+            }
+        }
+
         deleteAxis(axisName)
         addAxis(newAxis)
 
-        List<Column> newColumns = newAxis.columns
-        Map<Long, Long> columnMap = [:]
-        for (Column oldCol : oldColumns)
-        { // 1:1 map of old column ids to new column ids
-            columnMap[oldCol.id] = newColumns.find { Column newCol ->
-                newCol.valueThatMatches == oldCol.valueThatMatches
-            }.id
-        }
-
+        cells.clear()
         // change cell ids and put back into cube
-        for (Map.Entry<Set<Long>, T> cellMapEntry : cellMapCopy)
+        for (Map.Entry<Set<Long>, T> entry : cellMapCopy)
         {
-            Set<Long> coord = cellMapEntry.key
+            Set<Long> coord = entry.key
             // change coord to have existing ref ax value
             for (long oldCoordPart : coord)
             {
-                Long newCoordPart = columnMap[oldCoordPart]
-                if (newCoordPart) {
+                Long newCoordPart = oldToNewId[oldCoordPart]
+                if (newCoordPart)
+                {
                     coord.remove(oldCoordPart)
                     coord.add(newCoordPart)
                 }
             }
-            cells[coord] = (T)internValue(cellMapEntry.value)
+            
+            cells[coord] = (T) internValue(entry.value)
         }
+
+        // Eliminate orphans, where source axis (A, B, C, D, E) pointed to existing ref axis (A, C, E).
+        // Cells in columns B & D must be dropped!
+        Iterator<Set<Long>> i = cells.keySet().iterator()
+        while (i.hasNext())
+        {
+            Set<Long> ids = i.next()
+            Iterator<Long> j = ids.iterator()
+
+            while (j.hasNext())
+            {
+                long id = j.next()
+                if (getAxisFromColumnId(id) == null)
+                {
+                    i.remove()
+                }
+            }
+        }
+
+        // clearSha1() // called by other APIs [deleteAxis(), addAxis()]
     }
 
     /**